refactor: 요청에 대한 행동 클래스 분리
