+++ src/main/java/webserver/requestmapping/behavior/RequestBehavior.java
@@ -1,0 +1,9 @@
+package webserver.requestmapping.behavior;
+
+import java.io.InputStream;
+
+import http.request.RequestEntity;
+
+public interface RequestBehavior {
+    InputStream behave(RequestEntity requestEntity);
+}
