+++ src/test/java/webserver/requestmapping/FileExtensionTypeTest.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.requestmapping;
 
 import static org.assertj.core.api.Assertions.*;
